[WIP] coevolution of fitness predictors
